Dialer: Fix missing add to contacts option in callstats
We use the detailed callstat feature originally developed by AOKP
and fixed up by CM ported to 4.4.

During the merge to 4.4.1 we (as well CM) needed to remove for first
the loadercallback to check the status of the contact. Both teams (CM and Slim)
did not fix it till now but Google fixed same problem 2 month ago here:

https://github.com/SlimRoms/packages_apps_Dialer/commit/cce501cbde0ba0a53e1f458328bb96dff4dd6588

We follow this approach and adapt our DetailHeader code to work with
the Loadercallback and skip the binding of the contact and prepare
the option to add it as new if the current contact the user choosed
is not known (to be specific if this is a json encoded URI)

This fixes reported issue on xda

http://forum.xda-developers.com/showpost.php?p=49409411&postcount=3474

As well a small caching fix I added which got lost during the merge.

Change-Id: Ibe6d097357eb9b47bfc9dd1447084e1c72c4a318